Refactor vuex
