+++ src/login/index.ts
@@ -1,3 +1,4 @@
 import LoginRouter from "./login.router";
+import LoginStore from "./login.store";
 
-export { LoginRouter };
+export { LoginRouter, LoginStore };
