+++ src/main.ts
@@ -1,8 +1,8 @@
 import Vue from "vue";
 import App from "./App.vue";
-import store from "./store";
+import store from "@/plugins/vuex";
 import router from "@/plugins/router";
-import vuetify from "./plugins/vuetify";
+import vuetify from "@/plugins/vuetify";
 import vueCookie from "vue-cookies";
 import Vuelidate from "vuelidate";
 import VueApollo from "vue-apollo";
