+++ src/plugins/vuex.ts
@@ -1,10 +1,6 @@
 import Vue from "vue";
 import Vuex from "vuex";
-
-import posts from "@/store/modules/posts";
-import user from "@/store/modules/user";
-import likes from "@/store/modules/likes";
-import comment from "@/store/modules/comments";
+import { LoginStore } from "@/login";
 
 Vue.use(Vuex);
 
@@ -13,9 +9,6 @@ export default new Vuex.Store({
   mutations: {},
   actions: {},
   modules: {
-    user,
-    posts,
-    likes,
-    comment
+    LoginStore
   }
 });
